Move pkg/component to component to get past the validate-pkg check.
