+++ cmd/docker/docker.go
@@ -10,9 +10,9 @@ import (
 	"github.com/docker/docker/cli/command/commands"
 	cliflags "github.com/docker/docker/cli/flags"
 	"github.com/docker/docker/cliconfig"
+	"github.com/docker/docker/component"
+	compreg "github.com/docker/docker/component/registry"
 	"github.com/docker/docker/dockerversion"
-	"github.com/docker/docker/pkg/component"
-	compreg "github.com/docker/docker/pkg/component/registry"
 	"github.com/docker/docker/pkg/term"
 	"github.com/docker/docker/utils"
 	"github.com/spf13/cobra"
